requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.1.0
class-registry==2.1.2
colorama==0.4.6
filters==1.3.2
flake8==6.0.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
loguru==0.6.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
requests==2.28.2
six==1.16.0
tornado==6.2
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
win32-setctime==1.1.0
